Add files via upload
TikTok is one of the most popular social media apps today. It is famous for its short videos. People often use this app to watch entertaining and funny videos. Despite its entertaining content, this app is not liked by everyone. If you browse through his reviews, you'll find a mix of hate and support for TikTok around the world. So, let's analyze TikTok reviews to find out what people feel about the content of this app.